refactor: 컨벤션에 맞게 폴더 구조 정리
