+++ apps/host/src/entities/event-edit/types/event-edit.ts
@@ -1,0 +1,22 @@
+export interface FestivalDetail {
+  festivalId: number;
+  title: string;
+  location: string;
+  mainImageUrl: string;
+  activeCategoryIds: number[];
+  schedules: {
+    id: number;
+    festivalDate: string;
+    festivalTime: string;
+  }[];
+  stages: {
+    id: number;
+    title: string;
+    location?: string;
+  }[];
+}
+
+export interface UpdateFestivalResponse {
+  festivalId: number;
+  updatedAt: string;
+}
