+++ apps/host/src/pages/event-edit/event-edit.tsx
@@ -7,10 +7,11 @@ import { Loading } from '@amp/compositions';
 
 import EventForm from '@widgets/event-form/event-form';
 
-import { EVENT_EDIT_QUERY_OPTIONS } from '@entities/event-edit/apis/query';
-import { serializeUpdateFestivalFormData } from '@entities/event-edit/serialize-update-event-form-data';
-import { toEventEditInitialValues } from '@entities/event-edit/to-event-edit-initial-values';
-import { useFestivalUpdateMutation } from '@entities/event-edit/use-event-edit';
+import { serializeUpdateFestivalFormData } from '@features/event-edit/serialize-update-event-form-data';
+import { toEventEditInitialValues } from '@features/event-edit/to-event-edit-initial-values';
+import { useFestivalUpdateMutation } from '@features/event-edit/use-event-edit';
+
+import { EVENT_EDIT_QUERY_OPTIONS } from '@entities/event-edit/model/query-options';
 
 import { NAV_PATH } from '@shared/constants/path';
 import type { EventFormSubmitValues } from '@shared/types/event-form';
